requirements.txt
pip>=21.1.3
wheel>=0.36.2
pymongo~=3.12.0
future~=0.18.2
PyYAML~=5.4.1
requests~=2.26.0
yarg~=0.1.9
iso8601~=0.1.14
certifi~=2021.5.30
chardet~=4.0.0
docopt~=0.6.2
pipreqs~=0.4.10
urllib3~=1.26.6
six~=1.16.0
python-dateutil~=2.8.2
idna~=3.2
pycparser~=2.20
setuptools>=57.4.0
pythonnet>=2.5.1
pyserial~=3.5
minimalmodbus~=2.0.1
pandas~=1.4.3
numpy~=1.23.0
openpyxl~=3.0.7
opencv-python~=4.5.3
pyzbar~=0.1.8
ortools~=9.1.9490
scipy~=1.8.1
findpeaks~=2.4.0
ttkwidgets~=0.12.1